+++ inject/inject-server/src/main/scala/com/twitter/inject/server/Awaiter.scala
@@ -2,7 +2,7 @@ package com.twitter.inject.server
 
 import com.twitter.finagle.util.DefaultTimer
 import com.twitter.inject.Logging
-import com.twitter.util.{Await, Awaitable, Duration, Future}
+import com.twitter.util.{Await, Awaitable, Duration}
 import java.util.concurrent.CountDownLatch
 
 private[server] object Awaiter extends Logging {
@@ -16,25 +16,25 @@ private[server] object Awaiter extends Logging {
   def any(
     awaitables: Iterable[Awaitable[_]],
     period: Duration
-  ): Future[Unit] = {
-    info(
-      s"Awaiting ${awaitables.size} awaitables: \n${awaitables.map(_.getClass.getName).mkString("\n")}")
-    // Exit if ANY Awaitable is ready.
-    val latch = new CountDownLatch(1)
-    val task = DefaultTimer.schedule(period) {
-      if (awaitables.exists(Await.isReady)) {
-        awaitables.foreach { a: Awaitable[_] =>
-          // It is not expected that these Awaitables will exit here, but rather that the server
-          // will instead close thus triggering the Awaitables to also close. Therefore a "ready"
-          // Awaitable here is considered an error and logged as such.
-          if (Await.isReady(a)) error(s"${a.getClass.getName} awaitable has exited.")
+  ): Unit = {
+    if (awaitables.nonEmpty) {
+      info(s"Awaiting ${awaitables.size} awaitables: \n${awaitables
+        .map(_.getClass.getName)
+        .mkString("\n")}")
+      // Exit if ANY Awaitable is ready.
+      val latch = new CountDownLatch(1)
+      val task = DefaultTimer.schedule(period) {
+        if (awaitables.exists(Await.isReady)) {
+          awaitables.foreach { a: Awaitable[_] =>
+            if (Await.isReady(a)) warn(s"${a.getClass.getName} awaitable has exited.")
+          }
+          latch.countDown()
         }
-        latch.countDown()
       }
-    }
 
-    // We don't set a timeout because it should already be ready.
-    latch.await()
-    task.close()
+      // We don't set a timeout because it should already be ready.
+      latch.await()
+      Await.result(task.close(), period) // wait at most the period for the timer task to close
+    }
   }
 }